# lib/api-zod/src/generated/types/output.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ComfyUI Studio API — proxy and orchestration layer for a ComfyUI server
 * OpenAPI spec version: 0.1.0
 */
import type { OutputOutputType } from './outputOutputType';

export interface Output {
  id: number;
  jobId: number;
  filename: string;
  subfolder: string;
  outputType: OutputOutputType;
  /** Proxied URL to fetch the file from ComfyUI */
  comfyUrl: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  createdAt: Date;
}